pyproject.toml
[project]
name = "hospital-smart-assistant"
version = "0.1.0"
description = "Hospital smart assistant API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.134.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.40.0",
    "python-jose[cryptography]>=3.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.1",
    "ruff>=0.15.4",
    "fastapi-cli>=0.0.24",
]

[tool.ruff]
# Disesuaikan dengan Black formatter
line-length = 88

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["app/tests"]
